+++ src/decl/empty-decl.h
@@ -1,0 +1,1 @@
+static r_obj* list_drop_empty(r_obj* x);