+++ src/init.c
@@ -136,7 +136,7 @@ extern SEXP vctrs_chr_paste_prefix(SEXP, SEXP, SEXP);
 extern r_obj* vctrs_rank(r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* vctrs_integer64_proxy(r_obj*);
 extern r_obj* vctrs_integer64_restore(r_obj*);
-extern r_obj* vctrs_list_compact(r_obj*, r_obj*);
+extern r_obj* vctrs_list_drop_empty(r_obj*);
 
 
 // Maturing
@@ -291,7 +291,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_rank",                       (DL_FUNC) &vctrs_rank, 7},
   {"vctrs_integer64_proxy",            (DL_FUNC) &vctrs_integer64_proxy, 1},
   {"vctrs_integer64_restore",          (DL_FUNC) &vctrs_integer64_restore, 1},
-  {"vctrs_list_compact",               (DL_FUNC) &vctrs_list_compact, 2},
+  {"vctrs_list_drop_empty",            (DL_FUNC) &vctrs_list_drop_empty, 1},
   {NULL, NULL, 0}
 };